Rename code-mirror to codemirror & fix css editor
