+++ editor/js/editable-js.js
@@ -6,7 +6,7 @@ import * as mceUtils from "./editor-libs/mce-utils.js";
 import "../css/editor-libs/ui-fonts.css";
 import "../css/editor-libs/common.css";
 import "../css/editable-js-and-wat.css";
-import {initCodeEditor, getEditorContent, languageJavaScript} from "./editor-libs/code-mirror-editor.js";
+import {initCodeEditor, getEditorContent, languageJavaScript} from "./editor-libs/codemirror-editor.js";
 
 (function () {
   var codeBlock = document.getElementById("static-js");
