+++ editor/js/editor.js
@@ -3,7 +3,7 @@ import mceConsole from "./editor-libs/console.js";
 import * as mceEvents from "./editor-libs/events.js";
 import * as mceUtils from "./editor-libs/mce-utils.js";
 import * as tabby from "./editor-libs/tabby.js";
-import {getEditorContent} from "./editor-libs/code-mirror-editor.js";
+import {getEditorContent} from "./editor-libs/codemirror-editor.js";
 
 import "../css/editor-libs/ui-fonts.css";
 import "../css/editor-libs/common.css";
