+++ editor/tmpl/live-js-tmpl.html
@@ -5,7 +5,7 @@
     <meta http-equiv="x-ua-compatible" content="ie=edge" />
     <title>%title%</title>
     <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <link href="../../css/code-mirror.css%cache-buster%" rel="stylesheet" />
+    <link href="../../css/codemirror.css%cache-buster%" rel="stylesheet" />
     <link href="../../css/editor-js.css%cache-buster%" rel="stylesheet" />
     %example-css-src%
   </head>
@@ -32,7 +32,7 @@
 
     %example-js-src%
 
-    <script src="../../js/code-mirror.js%cache-buster%"></script>
+    <script src="../../js/codemirror.js%cache-buster%"></script>
     <script src="../../js/editor-js.js%cache-buster%"></script>
   </body>
 </html>
