+++ editor/tmpl/live-tabbed-tmpl.html
@@ -5,7 +5,7 @@
     <meta http-equiv="x-ua-compatible" content="ie=edge" />
     <title>%title%</title>
     <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <link href="../../css/code-mirror.css%cache-buster%" rel="stylesheet" />
+    <link href="../../css/codemirror.css%cache-buster%" rel="stylesheet" />
     <link href="../../css/editor-tabbed.css%cache-buster%" rel="stylesheet" />
   </head>
 
@@ -144,7 +144,7 @@
         }
       </script>
     </template>
-    <script src="../../js/code-mirror.js%cache-buster%"></script>
+    <script src="../../js/codemirror.js%cache-buster%"></script>
     <script src="../../js/editor-tabbed.js%cache-buster%"></script>
   </body>
 </html>
