+++ editor/tmpl/live-wat-tmpl.html
@@ -5,7 +5,7 @@
     <meta http-equiv="x-ua-compatible" content="ie=edge" />
     <title>%title%</title>
     <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <link href="../../css/code-mirror.css%cache-buster%" rel="stylesheet" />
+    <link href="../../css/codemirror.css%cache-buster%" rel="stylesheet" />
     <link href="../../css/editor-wat.css%cache-buster%" rel="stylesheet" />
   </head>
   <body>
@@ -70,7 +70,7 @@
       </section>
     </section>
 
-    <script src="../../js/code-mirror.js%cache-buster%"></script>
+    <script src="../../js/codemirror.js%cache-buster%"></script>
     <script src="../../js/editor-wat.js%cache-buster%"></script>
   </body>
 </html>
